% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSClassifier.R
\name{classify_nmibc}
\alias{classify_nmibc}
\title{Title}
\usage{
classify_nmibc(data, cor_cut = 0.1)
}
\arguments{
\item{data}{A dataframe where rows represent unique genes and columns
represent samples. Gene identifiers (rownames) can be provided as HUGO gene
symbols.RNA-seq data should be log-transformed,
for example using log2(normalized counts + 1).}

\item{cor_cut}{A numeric value specifying the minimum Pearson correlation
threshold for classification.Samples with a maximum correlation below this
threshold will remain unclassified, and their classification results will be
set to NA. Default value is 0.1.}
}
\value{
A dataframe containing classification results for each sample.
The `consensus_cluster` column provides the predicted consensus class label
for each sample. Additional columns include the Pearson correlation values
for each sample with each centroid, the `normalized_difference`
(a measure of how distinct the sample is from other classes),
and the `p_value` associated with the correlation to the nearest centroid.
Samples with a maximum correlation below the `cor_cut` threshold will have
NA in the `consensus_cluster` column.
}
\description{
Title
}
\examples{
# Example usage:
# `data` is a gene expression matrix:
data(tcgadat)
classify_nmibc(dejongA, cor_cut = 0.1)
}
